replace semi-public class_base with private _td_reftype
class_base is really only used to get the reftype of an item found in
the `$want` struct.  Really, it only has to return reftype, unless we're
in an old Perl where a regexp ref might be a SCALAR reftype under the
hood.

This commit removes it from exports and renames it to be more obviously
private.  It changes its return to just the reftype to use.

Nothing on CPAN seems to use class_base.  It is not documented.