defer wrapping leaves passed to all/any/none
Before this change, leaves (plain strings and numbers) passed to
all/any/none were converted to "real tests" with Test::Deep::wrap
during the all/any/none test's initialization.  This meant that
they were subject to the $LeafWrapper in place when any() was called,
rather than when the enclosing cmp_deeply was called.  This meant that
leaves inside and outside of all/any/none were treated differently.

Here's a test, not suitable for inclusion, that demonstrated the
problem:

  use strict;
  use warnings;

  use JSON;
  use JSON::Typist;
  use Test::Deep;
  use Test::Deep::JType;
  use Test::More;

  my $json = q<{"foo":1}>;
  my $data = JSON::Typist->new->apply_types(decode_json($json));

  jcmp_deeply($data->{foo}, 1);
  jcmp_deeply($data->{foo}, jnum(1));
  jcmp_deeply($data->{foo}, any(jnum(1)));
  jcmp_deeply($data->{foo}, any(1)); # <-- fails, 1 doesn't right wrapper

  done_testing;