add tests for correct timing of wrapping of all/any/none args
This commit moves tests for $LeafWrapper from shallow.t to a new test
file.

It also adds tests for the bug caused by all/any/none wrapping their
arguments at init time instead of at test time.  Without the fix, the
second set of tests fail because the $LeafWrapper in play for the calls
to any and all is shallow() and not str().