restore v5.12 compatibility
Funny story, here.  This test was added in 2019, but didn't make it into
the MANIFEST, so it didn't go out in the distribution, it only got run
on my machine -- where it was always running on v5.30-something.  When
the dist was converted to Dist::Zilla, the test file made it into the
build, exposing that it had secretly required v5.14!

Fixed!